refactor(core): split agent/mod.rs monolith into focused submodules
agent/mod.rs held a single impl<C: Channel> Agent<C> block of 58 methods
across 4120 lines — the largest production monolith in the workspace. Extract
cohesive concern groups into focused submodules, leaving mod.rs (now ~2000
lines) with the struct, constructors, and the main agent loop:

- shutdown.rs        — session-summary generation + orphaned tool-use flush
- subagent_commands.rs — /agent command handlers, sub-agent polling, and
                         spawn-context assembly (estimate_parts_size,
                         trim_parent_messages, sanitize_parent_messages)
- skill_reload.rs    — skill / instruction hot-reload
- config_reload.rs   — config overlay hot-reload
- hooks_dispatch.rs  — CwdChanged / FileChanged hook dispatch + the
                       McpManagerDispatch adapter

The three SideQuest eviction methods move into the existing sidequest.rs.

The shared impl block is now split across files, so 14 cross-module
entrypoints are bumped to pub(super); all other moved methods stay private.
No public API change. The trim_parent_messages test import is updated to the
new subagent_commands path. Remove the stale do-not-split TODO superseded by
this change.

Closes #4923
