+++ crates/zeph-core/src/agent/mod.rs
@@ -1,11 +1,6 @@
 // SPDX-FileCopyrightText: 2026 Andrei G <bug-ops>
 // SPDX-License-Identifier: MIT OR Apache-2.0
 
-// TODO(arch-revised-2026-04-26): agent/ module split deferred to epic/m49+/agent-split.
-// Hard prerequisite: 100% task_supervisor adoption (currently 10/31 raw tokio::spawn
-// sites in this directory). See arch-assessment-revised-2026-04-26T02-04-23.md PR 1
-// and PR 8. Do not split this file until PR 1 is merged and a /specs/ entry exists.
-
 mod acp_commands;
 mod agent_access_impl;
 pub(crate) mod agent_supervisor;
@@ -17,6 +12,7 @@ pub(crate) mod channel_impl;
 mod cocoon_cmd;
 mod command_context_impls;
 pub(super) mod compression_feedback;
+mod config_reload;
 mod context;
 mod context_impls;
 pub(crate) mod context_manager;
@@ -25,6 +21,7 @@ pub mod error;
 mod experiment_cmd;
 pub(crate) mod focus;
 mod heuristic_promotion;
+mod hooks_dispatch;
 mod index;
 mod learning;
 pub(crate) mod learning_engine;
@@ -52,11 +49,14 @@ mod scope_commands;
 pub mod session_config;
 mod session_digest;
 pub mod shadow_sentinel;
+mod shutdown;
 pub(crate) mod sidequest;
 mod skill_management;
+mod skill_reload;
 pub mod slash_commands;
 pub mod speculative;
 pub(crate) mod state;
+mod subagent_commands;
 pub(crate) mod task_injection;
 pub(crate) mod tool_execution;
 pub(crate) mod tool_orchestrator;
@@ -81,7 +81,7 @@ use zeph_llm::provider::{LlmProvider, Message, MessageMetadata, Role};
 use zeph_memory::TokenCounter;
 use zeph_memory::semantic::SemanticMemory;
 use zeph_skills::loader::Skill;
-use zeph_skills::matcher::{SkillMatcher, SkillMatcherBackend};
+use zeph_skills::matcher::SkillMatcherBackend;
 use zeph_skills::prompt::format_skills_prompt;
 use zeph_skills::registry::SkillRegistry;
 use zeph_tools::executor::{ErasedToolExecutor, ToolExecutor};
@@ -90,7 +90,7 @@ use tracing::Instrument as _;
 
 use crate::channel::Channel;
 use crate::config::Config;
-use crate::context::{ContextBudget, build_system_prompt};
+use crate::context::build_system_prompt;
 use zeph_common::text::estimate_tokens;
 
 use loop_event::LoopEvent;
@@ -347,521 +347,6 @@ impl<C: Channel> Agent<C> {
     #[must_use]
     pub fn into_channel(self) -> C {
         self.channel
-    }
-
-    /// Poll all active sub-agents for completed/failed/canceled results.
-    ///
-    /// Non-blocking: returns immediately with a list of `(task_id, result)` pairs
-    /// for agents that have finished. Each completed agent is removed from the manager.
-    #[tracing::instrument(name = "core.agent.poll_subagents", skip_all, level = "debug")]
-    pub async fn poll_subagents(&mut self) -> Vec<(String, String)> {
-        let Some(mgr) = &mut self.services.orchestration.subagent_manager else {
-            return vec![];
-        };
-
-        let finished: Vec<String> = mgr
-            .statuses()
-            .into_iter()
-            .filter_map(|(id, status)| {
-                if matches!(
-                    status.state,
-                    zeph_subagent::SubAgentState::Completed
-                        | zeph_subagent::SubAgentState::Failed
-                        | zeph_subagent::SubAgentState::Canceled
-                ) {
-                    Some(id)
-                } else {
-                    None
-                }
-            })
-            .collect();
-
-        let mut results = vec![];
-        for task_id in finished {
-            match mgr.collect(&task_id).await {
-                Ok(result) => results.push((task_id, result)),
-                Err(e) => {
-                    tracing::warn!(task_id, error = %e, "failed to collect sub-agent result");
-                }
-            }
-        }
-        results
-    }
-
-    /// Call the LLM to generate a structured session summary with a configurable timeout.
-    ///
-    /// Falls back to plain-text chat if structured output fails or times out. Returns `None` on
-    /// any failure, logging a warning — callers must treat `None` as "skip storage".
-    ///
-    /// Each LLM attempt is bounded by `shutdown_summary_timeout_secs`; in the worst case
-    /// (structured call times out and plain-text fallback also times out) this adds up to
-    /// `2 * shutdown_summary_timeout_secs` of shutdown latency.
-    async fn call_llm_for_session_summary(
-        &self,
-        chat_messages: &[Message],
-    ) -> Option<zeph_memory::StructuredSummary> {
-        let provider = self.resolve_background_provider(
-            &self.services.memory.compaction.shutdown_summary_provider,
-        );
-        let timeout_dur = std::time::Duration::from_secs(
-            self.services
-                .memory
-                .compaction
-                .shutdown_summary_timeout_secs,
-        );
-        match tokio::time::timeout(
-            timeout_dur,
-            provider.chat_typed_erased::<zeph_memory::StructuredSummary>(chat_messages),
-        )
-        .await
-        {
-            Ok(Ok(s)) => Some(s),
-            Ok(Err(e)) => {
-                tracing::warn!(
-                    "shutdown summary: structured LLM call failed, falling back to plain: {e:#}"
-                );
-                self.plain_text_summary_fallback(&provider, chat_messages, timeout_dur)
-                    .await
-            }
-            Err(_) => {
-                tracing::warn!(
-                    "shutdown summary: structured LLM call timed out after {}s, falling back to plain",
-                    self.services
-                        .memory
-                        .compaction
-                        .shutdown_summary_timeout_secs
-                );
-                self.plain_text_summary_fallback(&provider, chat_messages, timeout_dur)
-                    .await
-            }
-        }
-    }
-
-    async fn plain_text_summary_fallback(
-        &self,
-        provider: &zeph_llm::any::AnyProvider,
-        chat_messages: &[Message],
-        timeout_dur: std::time::Duration,
-    ) -> Option<zeph_memory::StructuredSummary> {
-        match tokio::time::timeout(timeout_dur, provider.chat(chat_messages)).await {
-            Ok(Ok(plain)) => Some(zeph_memory::StructuredSummary {
-                summary: plain,
-                key_facts: vec![],
-                entities: vec![],
-            }),
-            Ok(Err(e)) => {
-                tracing::warn!("shutdown summary: plain LLM fallback failed: {e:#}");
-                None
-            }
-            Err(_) => {
-                tracing::warn!("shutdown summary: plain LLM fallback timed out");
-                None
-            }
-        }
-    }
-
-    /// Persist tombstone `ToolResult` messages for any assistant `ToolUse` parts that were written
-    /// to the DB during this session but never paired with a `ToolResult` (e.g. because stdin
-    /// closed while tool execution was in progress). Without this the next session startup strips
-    /// those assistant messages and emits orphan warnings.
-    async fn flush_orphaned_tool_use_on_shutdown(&mut self) {
-        use zeph_llm::provider::{MessagePart, Role};
-
-        // Walk messages in reverse: if the last assistant message (ignoring any trailing
-        // system messages) has ToolUse parts and is NOT immediately followed by a user
-        // message whose ToolResult ids cover those ToolUse ids, persist tombstones.
-        let msgs = &self.msg.messages;
-        // Find last assistant message index.
-        let Some(asst_idx) = msgs.iter().rposition(|m| m.role == Role::Assistant) else {
-            return;
-        };
-        let asst_msg = &msgs[asst_idx];
-        let tool_use_ids: Vec<(&str, &str, &serde_json::Value)> = asst_msg
-            .parts
-            .iter()
-            .filter_map(|p| {
-                if let MessagePart::ToolUse { id, name, input } = p {
-                    Some((id.as_str(), name.as_str(), input))
-                } else {
-                    None
-                }
-            })
-            .collect();
-        if tool_use_ids.is_empty() {
-            return;
-        }
-
-        // Check whether a following user message already pairs all ToolUse ids.
-        let paired_ids: std::collections::HashSet<&str> = msgs
-            .get(asst_idx + 1..)
-            .into_iter()
-            .flatten()
-            .filter(|m| m.role == Role::User)
-            .flat_map(|m| m.parts.iter())
-            .filter_map(|p| {
-                if let MessagePart::ToolResult { tool_use_id, .. } = p {
-                    Some(tool_use_id.as_str())
-                } else {
-                    None
-                }
-            })
-            .collect();
-
-        let unpaired: Vec<zeph_llm::provider::ToolUseRequest> = tool_use_ids
-            .iter()
-            .filter(|(id, _, _)| !paired_ids.contains(*id))
-            .map(|(id, name, input)| zeph_llm::provider::ToolUseRequest {
-                id: (*id).to_owned(),
-                name: (*name).to_owned().into(),
-                input: (*input).clone(),
-            })
-            .collect();
-
-        if unpaired.is_empty() {
-            return;
-        }
-
-        tracing::info!(
-            count = unpaired.len(),
-            "shutdown: persisting tombstone ToolResults for unpaired in-flight tool calls"
-        );
-        self.persist_cancelled_tool_results(&unpaired).await;
-    }
-
-    /// Generate and store a lightweight session summary at shutdown when no hard compaction fired.
-    ///
-    /// Guards:
-    /// - `shutdown_summary` config must be enabled
-    /// - `conversation_id` must be set (memory must be attached)
-    /// - no existing session summary in the store (primary guard — resilient to failed Qdrant writes)
-    /// - at least `shutdown_summary_min_messages` user-turn messages in history
-    ///
-    /// All errors are logged as warnings and swallowed — shutdown must never fail.
-    async fn maybe_store_shutdown_summary(&mut self) {
-        if !self.services.memory.compaction.shutdown_summary {
-            return;
-        }
-        let Some(memory) = self.services.memory.persistence.memory.clone() else {
-            return;
-        };
-        let Some(conversation_id) = self.services.memory.persistence.conversation_id else {
-            return;
-        };
-
-        // Primary guard: check if a summary already exists (handles failed Qdrant writes too).
-        match memory.has_session_summary(conversation_id).await {
-            Ok(true) => {
-                tracing::debug!("shutdown summary: session already has a summary, skipping");
-                return;
-            }
-            Ok(false) => {}
-            Err(e) => {
-                tracing::warn!("shutdown summary: failed to check existing summary: {e:#}");
-                return;
-            }
-        }
-
-        // Count user-turn messages only (skip system prompt at index 0).
-        let user_count = self
-            .msg
-            .messages
-            .iter()
-            .skip(1)
-            .filter(|m| m.role == Role::User)
-            .count();
-        if user_count
-            < self
-                .services
-                .memory
-                .compaction
-                .shutdown_summary_min_messages
-        {
-            tracing::debug!(
-                user_count,
-                min = self
-                    .services
-                    .memory
-                    .compaction
-                    .shutdown_summary_min_messages,
-                "shutdown summary: too few user messages, skipping"
-            );
-            return;
-        }
-
-        // TUI status — send errors silently ignored (TUI may already be gone at shutdown).
-        let _ = self.channel.send_status("Saving session summary...").await;
-
-        // Collect last N messages (skip system prompt at index 0).
-        let max = self
-            .services
-            .memory
-            .compaction
-            .shutdown_summary_max_messages;
-        if max == 0 {
-            tracing::debug!("shutdown summary: max_messages=0, skipping");
-            return;
-        }
-        let non_system: Vec<_> = self.msg.messages.iter().skip(1).collect();
-        let slice = if non_system.len() > max {
-            &non_system[non_system.len() - max..]
-        } else {
-            &non_system[..]
-        };
-
-        let msgs_for_prompt: Vec<(zeph_memory::MessageId, String, String)> = slice
-            .iter()
-            .map(|m| {
-                let role = match m.role {
-                    Role::Assistant => "assistant".to_owned(),
-                    Role::System => "system".to_owned(),
-                    Role::User | _ => "user".to_owned(),
-                };
-                (zeph_memory::MessageId(0), role, m.content.clone())
-            })
-            .collect();
-
-        let prompt = zeph_memory::build_summarization_prompt(&msgs_for_prompt);
-        let chat_messages = vec![Message {
-            role: Role::User,
-            content: prompt,
-            parts: vec![],
-            metadata: MessageMetadata::default(),
-        }];
-
-        let Some(structured) = self.call_llm_for_session_summary(&chat_messages).await else {
-            let _ = self.channel.send_status("").await;
-            return;
-        };
-
-        if let Err(e) = memory
-            .store_shutdown_summary(conversation_id, &structured.summary, &structured.key_facts)
-            .await
-        {
-            tracing::warn!("shutdown summary: storage failed: {e:#}");
-        } else {
-            tracing::info!(
-                conversation_id = conversation_id.0,
-                "shutdown summary stored"
-            );
-        }
-
-        // Clear TUI status.
-        let _ = self.channel.send_status("").await;
-    }
-
-    /// Gracefully shut down the agent and persist state.
-    ///
-    /// Performs the following cleanup:
-    ///
-    /// 1. **Message persistence** — Deferred database writes (hide/summary operations)
-    ///    are flushed to memory or disk
-    /// 2. **Provider state** — LLM router state (e.g., Thompson sampling counters) is saved
-    ///    to the vault
-    /// 3. **Sub-agents** — All active sub-agent tasks are terminated
-    /// 4. **MCP servers** — All connected Model Context Protocol servers are shut down
-    /// 5. **Metrics finalization** — Compaction metrics and session metrics are recorded
-    /// 6. **Memory finalization** — Vector stores and semantic indices are flushed
-    /// 7. **Skill state** — Self-learning engine saves evolved skill definitions
-    ///
-    /// Call this before dropping the agent to ensure no data loss.
-    #[tracing::instrument(name = "core.agent.shutdown", skip_all, level = "debug")]
-    pub async fn shutdown(&mut self) {
-        let _ = self.channel.send_status("Shutting down...").await;
-
-        // CRIT-1: persist Thompson state accumulated during this session.
-        self.provider.save_router_state().await;
-
-        // Persist AdaptOrch Beta-arm table alongside Thompson state.
-        if let Some(ref advisor) = self.services.orchestration.topology_advisor
-            && let Err(e) = advisor.save()
-        {
-            tracing::warn!(error = %e, "adaptorch: failed to persist state");
-        }
-
-        if let Some(ref mut mgr) = self.services.orchestration.subagent_manager {
-            mgr.shutdown_all();
-        }
-
-        if let Some(ref manager) = self.services.mcp.manager {
-            manager.shutdown_all_shared().await;
-        }
-
-        // Finalize compaction trajectory: push the last open segment into the Vec.
-        // This segment would otherwise only be pushed when the next hard compaction fires,
-        // which never happens at session end.
-        if let Some(turns) = self.context_manager.turns_since_last_hard_compaction() {
-            self.update_metrics(|m| {
-                m.compaction_turns_after_hard.push(turns);
-            });
-            self.context_manager
-                .set_turns_since_last_hard_compaction(None);
-        }
-
-        if let Some(ref tx) = self.runtime.metrics.metrics_tx {
-            let m = tx.borrow();
-            if m.filter_applications > 0 {
-                #[allow(clippy::cast_precision_loss)]
-                let pct = if m.filter_raw_tokens > 0 {
-                    m.filter_saved_tokens as f64 / m.filter_raw_tokens as f64 * 100.0
-                } else {
-                    0.0
-                };
-                tracing::info!(
-                    raw_tokens = m.filter_raw_tokens,
-                    saved_tokens = m.filter_saved_tokens,
-                    applications = m.filter_applications,
-                    "tool output filtering saved ~{} tokens ({pct:.0}%)",
-                    m.filter_saved_tokens,
-                );
-            }
-            if m.compaction_hard_count > 0 {
-                tracing::info!(
-                    hard_compactions = m.compaction_hard_count,
-                    turns_after_hard = ?m.compaction_turns_after_hard,
-                    "hard compaction trajectory"
-                );
-            }
-        }
-
-        // Flush tombstone ToolResults for any assistant ToolUse that was persisted but never
-        // paired with a ToolResult (e.g. stdin EOF mid-execution). Without this the next session
-        // startup strips the orphaned ToolUse and emits warnings.
-        self.flush_orphaned_tool_use_on_shutdown().await;
-
-        // Signal the experiment CancellationToken first so the task can clean up gracefully,
-        // then abort the handle to guarantee it does not outlive the agent regardless.
-        if let Some(ref token) = self.services.experiments.cancel {
-            token.cancel();
-        }
-        if let Some(h) = self.services.experiments.handle.take() {
-            h.abort();
-        }
-
-        // Signal cooperative cancellation to the graph-extraction background task before the
-        // hard abort below. This lets the task exit at a clean checkpoint (e.g. after the
-        // community-refresh select arm fires) rather than being cut mid-write.
-        if let Some(memory) = self.services.memory.persistence.memory.as_ref() {
-            memory.cancel_graph_extraction();
-        }
-
-        // Forcibly abort in-flight Enrichment and Telemetry tasks tracked by the supervisor.
-        self.runtime.lifecycle.supervisor.abort_all();
-
-        // Abort background task handles not tracked by BackgroundSupervisor.
-        // Per the Await Discipline rule, fire-and-forget handles must be aborted on shutdown.
-        if let Some(h) = self.services.compression.pending_task_goal.take() {
-            h.abort();
-        }
-        if let Some(h) = self.services.compression.pending_sidequest_result.take() {
-            h.abort();
-        }
-        if let Some(h) = self.services.compression.pending_subgoal.take() {
-            h.abort();
-        }
-
-        // Abort learning tasks (JoinSet detached at turn boundaries but not on shutdown).
-        self.services.learning_engine.learning_tasks.abort_all();
-
-        // Await the AutoSkill trace extraction task so it is not silently dropped.
-        // Bounded to avoid hanging shutdown when the LLM call inside the task stalls.
-        if let Some(h) = self.services.learning_engine.trace_extraction_handle.take() {
-            let deadline = std::time::Duration::from_mins(2);
-            match tokio::time::timeout(deadline, h).await {
-                Ok(Ok(())) => {}
-                Ok(Err(e)) => tracing::warn!("trace_extraction: task panicked at shutdown: {e}"),
-                Err(_) => tracing::warn!(
-                    "trace_extraction: timed out at shutdown ({}s), aborting",
-                    deadline.as_secs()
-                ),
-            }
-        }
-
-        // Abort the heuristic promotion loop (periodic task; abort is safe because
-        // promotion_already_evaluated ensures idempotent retry on next startup).
-        if let Some(h) = self
-            .services
-            .learning_engine
-            .heuristic_promotion_handle
-            .take()
-        {
-            h.abort();
-        }
-
-        // Drain pending shadow sentinel DB writes before final teardown.
-        if let Some(ref sentinel) = self.services.security.shadow_sentinel {
-            sentinel.drain_pending().await;
-        }
-
-        // Allow cancelled tasks to release their HTTP connections before the summary LLM call.
-        // abort_all() posts cancellation signals but does not drain tasks; aborted futures only
-        // observe cancellation at their next .await point. Without yielding here the summary
-        // call races in-flight enrichment HTTP connections for the same API rate-limit budget.
-        for _ in 0..4 {
-            tokio::task::yield_now().await;
-        }
-
-        self.maybe_store_shutdown_summary().await;
-        self.maybe_store_session_digest().await;
-
-        tracing::info!("agent shutdown complete");
-    }
-
-    /// Run the chat loop, receiving messages via the channel until EOF or shutdown.
-    ///
-    /// # Errors
-    ///
-    /// Returns an error if channel I/O or LLM communication fails.
-    /// Refresh sub-agent metrics snapshot for the TUI metrics panel.
-    fn refresh_subagent_metrics(&mut self) {
-        let Some(ref mgr) = self.services.orchestration.subagent_manager else {
-            return;
-        };
-        let sub_agent_metrics: Vec<crate::metrics::SubAgentMetrics> = mgr
-            .statuses()
-            .into_iter()
-            .map(|(id, s)| {
-                let def = mgr.agents_def(&id);
-                crate::metrics::SubAgentMetrics {
-                    name: def.map_or_else(|| id[..8.min(id.len())].to_owned(), |d| d.name.clone()),
-                    id: id.clone(),
-                    state: format!("{:?}", s.state).to_lowercase(),
-                    turns_used: s.turns_used,
-                    max_turns: def.map_or(20, |d| d.permissions.max_turns),
-                    background: def.is_some_and(|d| d.permissions.background),
-                    elapsed_secs: s.started_at.elapsed().as_secs(),
-                    permission_mode: def.map_or_else(String::new, |d| {
-                        use zeph_subagent::def::PermissionMode;
-                        match d.permissions.permission_mode {
-                            PermissionMode::AcceptEdits => "accept_edits".into(),
-                            PermissionMode::DontAsk => "dont_ask".into(),
-                            PermissionMode::BypassPermissions => "bypass_permissions".into(),
-                            PermissionMode::Plan => "plan".into(),
-                            _ => String::new(),
-                        }
-                    }),
-                    transcript_dir: mgr
-                        .agent_transcript_dir(&id)
-                        .map(|p| p.to_string_lossy().into_owned()),
-                }
-            })
-            .collect();
-        self.update_metrics(|m| m.sub_agents = sub_agent_metrics);
-    }
-
-    /// Non-blocking poll: notify the user when background sub-agents complete.
-    async fn notify_completed_subagents(&mut self) -> Result<(), error::AgentError> {
-        let completed = self.poll_subagents().await;
-        for (task_id, result) in completed {
-            let notice = if result.is_empty() {
-                format!("[sub-agent {id}] completed (no output)", id = &task_id[..8])
-            } else {
-                format!("[sub-agent {id}] completed:\n{result}", id = &task_id[..8])
-            };
-            if let Err(e) = self.channel.send(&notice).await {
-                tracing::warn!(error = %e, "failed to send sub-agent completion notice");
-            }
-        }
-        Ok(())
     }
 
     /// Run the agent main loop.
@@ -2335,1365 +1820,6 @@ impl<C: Channel> Agent<C> {
         parts
     }
 
-    /// Poll a sub-agent until it reaches a terminal state, bridging secret requests to the
-    /// channel. Returns a human-readable status string and success flag suitable for
-    /// sending to the user and emitting lifecycle events.
-    async fn poll_subagent_until_done(
-        &mut self,
-        task_id: &str,
-        label: &str,
-    ) -> Option<(String, bool)> {
-        use zeph_subagent::SubAgentState;
-        let result = loop {
-            tokio::time::sleep(std::time::Duration::from_millis(500)).await;
-
-            // Bridge secret requests from sub-agent to channel.confirm().
-            // Fetch the pending request first, then release the borrow before
-            // calling channel.confirm() (which requires &mut self).
-            #[allow(clippy::redundant_closure_for_method_calls)]
-            let pending = self
-                .services
-                .orchestration
-                .subagent_manager
-                .as_mut()
-                .and_then(|m| m.try_recv_secret_request());
-            if let Some((req_task_id, req)) = pending {
-                // req.secret_key is pre-validated to [a-zA-Z0-9_-] in manager.rs
-                // (SEC-P1-02), so it is safe to embed in the prompt string.
-                let confirm_prompt = format!(
-                    "Sub-agent requests secret '{}'. Allow?",
-                    crate::text::truncate_to_chars(&req.secret_key, 100)
-                );
-                let approved = self.channel.confirm(&confirm_prompt).await.unwrap_or(false);
-                if let Some(mgr) = self.services.orchestration.subagent_manager.as_mut() {
-                    if approved {
-                        let ttl = std::time::Duration::from_mins(5);
-                        let key = req.secret_key.clone();
-                        if mgr.approve_secret(&req_task_id, &key, ttl).is_ok() {
-                            let _ = mgr.deliver_secret(&req_task_id, key);
-                        }
-                    } else {
-                        let _ = mgr.deny_secret(&req_task_id);
-                    }
-                }
-            }
-
-            let mgr = self.services.orchestration.subagent_manager.as_ref()?;
-            let statuses = mgr.statuses();
-            let Some((_, status)) = statuses.iter().find(|(id, _)| id == task_id) else {
-                break (format!("{label} completed (no status available)."), true);
-            };
-            match status.state {
-                SubAgentState::Completed => {
-                    let msg = status.last_message.clone().unwrap_or_else(|| "done".into());
-                    break (format!("{label} completed: {msg}"), true);
-                }
-                SubAgentState::Failed => {
-                    let msg = status
-                        .last_message
-                        .clone()
-                        .unwrap_or_else(|| "unknown error".into());
-                    break (format!("{label} failed: {msg}"), false);
-                }
-                SubAgentState::Canceled => {
-                    break (format!("{label} was cancelled."), false);
-                }
-                _ => {
-                    let _ = self
-                        .channel
-                        .send_status(&format!(
-                            "{label}: turn {}/{}",
-                            status.turns_used,
-                            self.services
-                                .orchestration
-                                .subagent_manager
-                                .as_ref()
-                                .and_then(|m| m.agents_def(task_id))
-                                .map_or(20, |d| d.permissions.max_turns)
-                        ))
-                        .await;
-                }
-            }
-        };
-        Some(result)
-    }
-
-    /// Resolve a unique full `task_id` from a prefix. Returns `None` if the manager is absent,
-    /// `Some(Err(msg))` on ambiguity/not-found, `Some(Ok(full_id))` on success.
-    fn resolve_agent_id_prefix(&mut self, prefix: &str) -> Option<Result<String, String>> {
-        let mgr = self.services.orchestration.subagent_manager.as_mut()?;
-        let full_ids: Vec<String> = mgr
-            .statuses()
-            .into_iter()
-            .map(|(tid, _)| tid)
-            .filter(|tid| tid.starts_with(prefix))
-            .collect();
-        Some(match full_ids.as_slice() {
-            [] => Err(format!("No sub-agent with id prefix '{prefix}'")),
-            [fid] => Ok(fid.clone()),
-            _ => Err(format!(
-                "Ambiguous id prefix '{prefix}': matches {} agents",
-                full_ids.len()
-            )),
-        })
-    }
-
-    fn handle_agent_list(&self) -> Option<String> {
-        use std::fmt::Write as _;
-        let mgr = self.services.orchestration.subagent_manager.as_ref()?;
-        let defs = mgr.definitions();
-        if defs.is_empty() {
-            return Some("No sub-agent definitions found.".into());
-        }
-        let mut out = String::from("Available sub-agents:\n");
-        for d in defs {
-            let memory_label = match d.memory {
-                Some(zeph_subagent::MemoryScope::User) => " [memory:user]",
-                Some(zeph_subagent::MemoryScope::Project) => " [memory:project]",
-                Some(zeph_subagent::MemoryScope::Local) => " [memory:local]",
-                Some(_) => " [memory:unknown]",
-                None => "",
-            };
-            if let Some(ref src) = d.source {
-                let _ = writeln!(
-                    out,
-                    "  {}{} — {} ({})",
-                    d.name, memory_label, d.description, src
-                );
-            } else {
-                let _ = writeln!(out, "  {}{} — {}", d.name, memory_label, d.description);
-            }
-        }
-        Some(out)
-    }
-
-    fn handle_agent_status(&self) -> Option<String> {
-        use std::fmt::Write as _;
-        let mgr = self.services.orchestration.subagent_manager.as_ref()?;
-        let statuses = mgr.statuses();
-        if statuses.is_empty() {
-            return Some("No active sub-agents.".into());
-        }
-        let mut out = String::from("Active sub-agents:\n");
-        for (id, s) in &statuses {
-            let state = format!("{:?}", s.state).to_lowercase();
-            let elapsed = s.started_at.elapsed().as_secs();
-            let _ = writeln!(
-                out,
-                "  [{short}] {state}  turns={t}  elapsed={elapsed}s  {msg}",
-                short = &id[..8.min(id.len())],
-                t = s.turns_used,
-                msg = s.last_message.as_deref().unwrap_or(""),
-            );
-            // Show memory directory path for agents with memory enabled.
-            if let Some(def) = mgr.agents_def(id)
-                && let Some(scope) = def.memory
-                && let Ok(dir) = zeph_subagent::memory::resolve_memory_dir(scope, &def.name)
-            {
-                let _ = writeln!(out, "       memory: {}", dir.display());
-            }
-        }
-        Some(out)
-    }
-
-    fn handle_agent_approve(&mut self, id: &str) -> Option<String> {
-        let full_id = match self.resolve_agent_id_prefix(id)? {
-            Ok(fid) => fid,
-            Err(msg) => return Some(msg),
-        };
-        let mgr = self.services.orchestration.subagent_manager.as_mut()?;
-        if let Some((tid, req)) = mgr.try_recv_secret_request()
-            && tid == full_id
-        {
-            let key = req.secret_key.clone();
-            let ttl = std::time::Duration::from_mins(5);
-            if let Err(e) = mgr.approve_secret(&full_id, &key, ttl) {
-                return Some(format!("Approve failed: {e}"));
-            }
-            if let Err(e) = mgr.deliver_secret(&full_id, key.clone()) {
-                return Some(format!("Secret delivery failed: {e}"));
-            }
-            return Some(format!("Secret '{key}' approved for sub-agent {full_id}."));
-        }
-        Some(format!(
-            "No pending secret request for sub-agent '{full_id}'."
-        ))
-    }
-
-    fn handle_agent_deny(&mut self, id: &str) -> Option<String> {
-        let full_id = match self.resolve_agent_id_prefix(id)? {
-            Ok(fid) => fid,
-            Err(msg) => return Some(msg),
-        };
-        let mgr = self.services.orchestration.subagent_manager.as_mut()?;
-        match mgr.deny_secret(&full_id) {
-            Ok(()) => Some(format!("Secret request denied for sub-agent '{full_id}'.")),
-            Err(e) => Some(format!("Deny failed: {e}")),
-        }
-    }
-
-    async fn handle_agent_command(&mut self, cmd: zeph_subagent::AgentCommand) -> Option<String> {
-        use zeph_subagent::AgentCommand;
-
-        match cmd {
-            AgentCommand::List => self.handle_agent_list(),
-            AgentCommand::Background { name, prompt } => {
-                self.handle_agent_background(&name, &prompt)
-            }
-            AgentCommand::Spawn { name, prompt }
-            | AgentCommand::Mention {
-                agent: name,
-                prompt,
-            } => self.handle_agent_spawn_foreground(&name, &prompt).await,
-            AgentCommand::Status => self.handle_agent_status(),
-            AgentCommand::Cancel { id } => self.handle_agent_cancel(&id),
-            AgentCommand::Approve { id } => self.handle_agent_approve(&id),
-            AgentCommand::Deny { id } => self.handle_agent_deny(&id),
-            AgentCommand::Resume { id, prompt } => self.handle_agent_resume(&id, &prompt).await,
-            _ => None,
-        }
-    }
-
-    /// Return the sub-agent definitions section formatted for the `/agents` fleet view.
-    ///
-    /// Produces a "Sub-agents:" header followed by one line per definition.
-    /// Returns an empty string when no sub-agent manager is configured.
-    pub(crate) fn handle_agents_definitions_list(&self) -> String {
-        use std::fmt::Write as _;
-
-        let Some(mgr) = self.services.orchestration.subagent_manager.as_ref() else {
-            return String::new();
-        };
-        let defs = mgr.definitions();
-        if defs.is_empty() {
-            return String::new();
-        }
-        let mut out = String::from("Sub-agents:\n");
-        for d in defs {
-            let memory_label = match d.memory {
-                Some(zeph_subagent::MemoryScope::User) => " [memory:user]",
-                Some(zeph_subagent::MemoryScope::Project) => " [memory:project]",
-                Some(zeph_subagent::MemoryScope::Local) => " [memory:local]",
-                Some(_) => " [memory:unknown]",
-                None => "",
-            };
-            if let Some(ref src) = d.source {
-                let _ = writeln!(
-                    out,
-                    "  {}{} — {} ({})",
-                    d.name, memory_label, d.description, src
-                );
-            } else {
-                let _ = writeln!(out, "  {}{} — {}", d.name, memory_label, d.description);
-            }
-        }
-        out
-    }
-
-    /// Execute an `/agents` CRUD subcommand and return a formatted string.
-    ///
-    /// Handles `show`, `create`, `edit`, `delete` (the `list` case is handled by
-    /// [`handle_agents_definitions_list`] and never reaches this method).
-    pub(crate) fn handle_agents_crud(&mut self, cmd: zeph_subagent::AgentsCommand) -> String {
-        use zeph_subagent::AgentsCommand;
-
-        let Some(mgr) = self.services.orchestration.subagent_manager.as_ref() else {
-            return "Sub-agent manager is not available.".to_owned();
-        };
-
-        match cmd {
-            AgentsCommand::List => self.handle_agents_definitions_list(),
-            AgentsCommand::Show { name } => {
-                match mgr.definitions().iter().find(|d| d.name == name) {
-                    Some(d) => format!(
-                        "Agent: {}\nDescription: {}\nSource: {}\n",
-                        d.name,
-                        d.description,
-                        d.source.as_deref().unwrap_or("unknown"),
-                    ),
-                    None => format!("No sub-agent definition named '{name}'."),
-                }
-            }
-            AgentsCommand::Create { name } => {
-                format!(
-                    "To create a sub-agent definition, create a file at `.zeph/agents/{name}.md`.\n\
-                     See the sub-agent documentation for the required frontmatter."
-                )
-            }
-            AgentsCommand::Edit { name } => {
-                format!("To edit '{name}', open its definition file in `.zeph/agents/{name}.md`.")
-            }
-            AgentsCommand::Delete { name } => {
-                format!("To delete '{name}', remove the file `.zeph/agents/{name}.md`.")
-            }
-            _ => "Unknown agents command.".to_owned(),
-        }
-    }
-
-    fn handle_agent_background(&mut self, name: &str, prompt: &str) -> Option<String> {
-        let provider = self.provider.clone();
-        let tool_executor = Arc::clone(&self.tool_executor);
-        let skills = self.filtered_skills_for(name);
-        let cfg = self.services.orchestration.subagent_config.clone();
-        let spawn_ctx = self.build_spawn_context(&cfg);
-        let mgr = self.services.orchestration.subagent_manager.as_mut()?;
-        match mgr.spawn(
-            name,
-            prompt,
-            provider,
-            tool_executor,
-            skills,
-            &cfg,
-            spawn_ctx,
-        ) {
-            Ok(id) => Some(format!(
-                "Sub-agent '{name}' started in background (id: {short})",
-                short = &id[..8.min(id.len())]
-            )),
-            Err(e) => Some(format!("Failed to spawn sub-agent: {e}")),
-        }
-    }
-
-    async fn handle_agent_spawn_foreground(&mut self, name: &str, prompt: &str) -> Option<String> {
-        let provider = self.provider.clone();
-        let tool_executor = Arc::clone(&self.tool_executor);
-        let skills = self.filtered_skills_for(name);
-        let cfg = self.services.orchestration.subagent_config.clone();
-        let spawn_ctx = self.build_spawn_context(&cfg);
-        let mgr = self.services.orchestration.subagent_manager.as_mut()?;
-        let task_id = match mgr.spawn(
-            name,
-            prompt,
-            provider,
-            tool_executor,
-            skills,
-            &cfg,
-            spawn_ctx,
-        ) {
-            Ok(id) => id,
-            Err(e) => return Some(format!("Failed to spawn sub-agent: {e}")),
-        };
-        let short = task_id[..8.min(task_id.len())].to_owned();
-        let _ = self
-            .channel
-            .send(&format!("Sub-agent '{name}' running... (id: {short})"))
-            .await;
-        let _ = self
-            .channel
-            .notify_foreground_subagent_started(&task_id, name)
-            .await;
-        let label = format!("Sub-agent '{name}'");
-        let Some((result, success)) = self.poll_subagent_until_done(&task_id, &label).await else {
-            // Manager was dropped mid-poll; emit completed(false) so TUI does not stay stuck.
-            let _ = self
-                .channel
-                .notify_foreground_subagent_completed(&task_id, name, false)
-                .await;
-            return None;
-        };
-        let _ = self
-            .channel
-            .notify_foreground_subagent_completed(&task_id, name, success)
-            .await;
-        Some(result)
-    }
-
-    fn handle_agent_cancel(&mut self, id: &str) -> Option<String> {
-        let mgr = self.services.orchestration.subagent_manager.as_mut()?;
-        // Accept prefix match on task_id.
-        let ids: Vec<String> = mgr
-            .statuses()
-            .into_iter()
-            .map(|(task_id, _)| task_id)
-            .filter(|task_id| task_id.starts_with(id))
-            .collect();
-        match ids.as_slice() {
-            [] => Some(format!("No sub-agent with id prefix '{id}'")),
-            [full_id] => {
-                let full_id = full_id.clone();
-                match mgr.cancel(&full_id) {
-                    Ok(()) => Some(format!("Cancelled sub-agent {full_id}.")),
-                    Err(e) => Some(format!("Cancel failed: {e}")),
-                }
-            }
-            _ => Some(format!(
-                "Ambiguous id prefix '{id}': matches {} agents",
-                ids.len()
-            )),
-        }
-    }
-
-    async fn handle_agent_resume(&mut self, id: &str, prompt: &str) -> Option<String> {
-        let cfg = self.services.orchestration.subagent_config.clone();
-        // Resolve definition name from transcript meta before spawning so we can
-        // look up skills by definition name rather than the UUID prefix (S1 fix).
-        let def_name = {
-            let mgr = self.services.orchestration.subagent_manager.as_ref()?;
-            match mgr.def_name_for_resume(id, &cfg) {
-                Ok(name) => name,
-                Err(e) => return Some(format!("Failed to resume sub-agent: {e}")),
-            }
-        };
-        let skills = self.filtered_skills_for(&def_name);
-        let provider = self.provider.clone();
-        let tool_executor = Arc::clone(&self.tool_executor);
-        let mgr = self.services.orchestration.subagent_manager.as_mut()?;
-        let (task_id, _) = match mgr.resume(id, prompt, provider, tool_executor, skills, &cfg, None)
-        {
-            Ok(pair) => pair,
-            Err(e) => return Some(format!("Failed to resume sub-agent: {e}")),
-        };
-        let short = task_id[..8.min(task_id.len())].to_owned();
-        let _ = self
-            .channel
-            .send(&format!("Resuming sub-agent '{id}'... (new id: {short})"))
-            .await;
-        let _ = self
-            .channel
-            .notify_foreground_subagent_started(&task_id, &def_name)
-            .await;
-        let Some((result, success)) = self
-            .poll_subagent_until_done(&task_id, "Resumed sub-agent")
-            .await
-        else {
-            // Manager was dropped mid-poll; emit completed(false) so TUI does not stay stuck.
-            let _ = self
-                .channel
-                .notify_foreground_subagent_completed(&task_id, &def_name, false)
-                .await;
-            return None;
-        };
-        let _ = self
-            .channel
-            .notify_foreground_subagent_completed(&task_id, &def_name, success)
-            .await;
-        Some(result)
-    }
-
-    fn filtered_skills_for(&self, agent_name: &str) -> Option<Vec<String>> {
-        let mgr = self.services.orchestration.subagent_manager.as_ref()?;
-        let def = mgr.definitions().iter().find(|d| d.name == agent_name)?;
-        let reg = self.services.skill.registry.read();
-        match zeph_subagent::filter_skills(&reg, &def.skills) {
-            Ok(skills) => {
-                let bodies: Vec<String> = skills.into_iter().map(|s| s.body.clone()).collect();
-                if bodies.is_empty() {
-                    None
-                } else {
-                    Some(bodies)
-                }
-            }
-            Err(e) => {
-                tracing::warn!(error = %e, "skill filtering failed for sub-agent");
-                None
-            }
-        }
-    }
-
-    /// Build a `SpawnContext` from current agent state for sub-agent spawning.
-    fn build_spawn_context(
-        &self,
-        cfg: &zeph_config::SubAgentConfig,
-    ) -> zeph_subagent::SpawnContext {
-        zeph_subagent::SpawnContext {
-            parent_messages: self.extract_parent_messages(cfg),
-            parent_cancel: Some(self.runtime.lifecycle.cancel_token.clone()),
-            parent_provider_name: {
-                let name = &self.runtime.config.active_provider_name;
-                if name.is_empty() {
-                    None
-                } else {
-                    Some(name.clone())
-                }
-            },
-            spawn_depth: self.runtime.config.spawn_depth,
-            mcp_tool_names: self.extract_mcp_tool_names(),
-            // F3 spec 050 §4: propagate seeded score when parent is >= Elevated.
-            seed_trajectory_score: {
-                let child = self.services.security.trajectory.spawn_child();
-                let score = child.score_now();
-                if score > 0.0 { Some(score) } else { None }
-            },
-            content_isolation: self.runtime.config.security.content_isolation.clone(),
-            orchestrator_name: Some("zeph".to_owned()),
-            orchestrator_role: Some("orchestrator".to_owned()),
-            session_mcp_servers: Vec::new(),
-            // Constraint propagation (#3993): populated by orchestration layer when spawning
-            // with explicit trust/tool restrictions. Top-level agent sessions leave these None.
-            ..Default::default()
-        }
-    }
-
-    /// Extract recent parent messages for history propagation (Section 5.7 in spec).
-    ///
-    /// Filters system messages, applies `context_window_turns` and `max_parent_messages` caps,
-    /// applies a 25% context window cap using a 4-chars-per-token heuristic, prunes orphaned
-    /// `ToolUse`/`ToolResult` pairs at the slice boundary, and optionally sanitizes text parts
-    /// through the IPI pipeline according to `parent_context_policy`.
-    fn extract_parent_messages(
-        &self,
-        config: &zeph_config::SubAgentConfig,
-    ) -> Vec<zeph_llm::provider::Message> {
-        use zeph_config::ParentContextPolicy;
-        use zeph_llm::provider::Role;
-
-        if config.parent_context_policy == ParentContextPolicy::None
-            || config.context_window_turns == 0
-        {
-            return Vec::new();
-        }
-
-        let non_system: Vec<_> = self
-            .msg
-            .messages
-            .iter()
-            .filter(|m| m.role != Role::System)
-            .cloned()
-            .collect();
-
-        let take_count = config
-            .context_window_turns
-            .saturating_mul(2)
-            .min(config.max_parent_messages);
-        let start = non_system.len().saturating_sub(take_count);
-        let mut msgs = non_system[start..].to_vec();
-
-        // Cap at 25% of model context window and prune orphaned tool pairs.
-        let max_chars = 128_000usize / 4;
-        let requested = msgs.len();
-        trim_parent_messages(&mut msgs, max_chars);
-        if msgs.len() < requested {
-            tracing::info!(
-                kept = msgs.len(),
-                requested,
-                "[subagent] truncated parent history due to token budget or orphan pruning"
-            );
-        }
-
-        if config.parent_context_policy == ParentContextPolicy::InheritSanitized {
-            use zeph_sanitizer::{ContentSource, ContentSourceKind};
-            let source =
-                ContentSource::new(ContentSourceKind::A2aMessage).with_identifier("parent_history");
-            msgs = sanitize_parent_messages(msgs, &self.services.security.sanitizer, &source);
-        }
-
-        msgs
-    }
-
-    /// Extract MCP tool names from the tool executor for diagnostic annotation.
-    fn extract_mcp_tool_names(&self) -> Vec<String> {
-        self.tool_executor
-            .tool_definitions_erased()
-            .into_iter()
-            .filter(|t| t.id.starts_with("mcp_"))
-            .map(|t| t.id.to_string())
-            .collect()
-    }
-
-    /// Classify a skill directory's source kind using on-disk markers and the bundled allowlist.
-    ///
-    /// Must be called from a blocking context (uses synchronous FS I/O).
-    fn classify_source_kind(
-        skill_dir: &std::path::Path,
-        managed_dir: Option<&std::path::PathBuf>,
-        bundled_names: &std::collections::HashSet<String>,
-    ) -> zeph_memory::store::SourceKind {
-        if managed_dir.is_some_and(|d| skill_dir.starts_with(d)) {
-            let skill_name = skill_dir.file_name().and_then(|n| n.to_str()).unwrap_or("");
-            let has_marker = skill_dir.join(".bundled").exists();
-            if has_marker && bundled_names.contains(skill_name) {
-                zeph_memory::store::SourceKind::Bundled
-            } else {
-                if has_marker {
-                    tracing::warn!(
-                        skill = %skill_name,
-                        "skill has .bundled marker but is not in the bundled skill \
-                         allowlist — classifying as Hub"
-                    );
-                }
-                zeph_memory::store::SourceKind::Hub
-            }
-        } else {
-            zeph_memory::store::SourceKind::Local
-        }
-    }
-
-    /// Update trust DB records for all reloaded skills.
-    async fn update_trust_for_reloaded_skills(
-        &mut self,
-        all_meta: &[zeph_skills::loader::SkillMeta],
-    ) {
-        // Clone Arc before any .await so no &self fields are held across suspension points.
-        let memory = self.services.memory.persistence.memory.clone();
-        let Some(memory) = memory else {
-            return;
-        };
-        let trust_cfg = self.services.skill.trust_config.clone();
-        let managed_dir = self.services.skill.managed_dir.clone();
-        let bundled_names: std::collections::HashSet<String> =
-            zeph_skills::bundled_skill_names().into_iter().collect();
-        for meta in all_meta {
-            // Compute hash and classify source_kind in spawn_blocking — both are blocking FS calls
-            // (.bundled marker .exists() and compute_skill_hash both do std::fs I/O).
-            let skill_dir = meta.skill_dir.clone();
-            let managed_dir_ref = managed_dir.clone();
-            let bundled_names_ref = bundled_names.clone();
-            let fs_result: Option<(String, zeph_memory::store::SourceKind)> =
-                tokio::task::spawn_blocking(move || {
-                    let hash = zeph_skills::compute_skill_hash(&skill_dir).ok()?;
-                    let source_kind = Self::classify_source_kind(
-                        &skill_dir,
-                        managed_dir_ref.as_ref(),
-                        &bundled_names_ref,
-                    );
-                    Some((hash, source_kind))
-                })
-                .await
-                .unwrap_or(None);
-
-            let Some((current_hash, source_kind)) = fs_result else {
-                tracing::warn!("failed to compute hash for '{}'", meta.name);
-                continue;
-            };
-            let initial_level = match source_kind {
-                zeph_memory::store::SourceKind::Bundled => &trust_cfg.bundled_level,
-                zeph_memory::store::SourceKind::Local | zeph_memory::store::SourceKind::File => {
-                    &trust_cfg.local_level
-                }
-                _ => &trust_cfg.default_level,
-            };
-            let existing = memory
-                .sqlite()
-                .load_skill_trust(&meta.name)
-                .await
-                .ok()
-                .flatten();
-            let trust_level = if let Some(ref row) = existing {
-                if row.blake3_hash != current_hash {
-                    trust_cfg.hash_mismatch_level
-                } else if row.source_kind != source_kind {
-                    // source_kind changed (e.g., hub → bundled on upgrade).
-                    // Never override an explicit operator block. For active trust levels,
-                    // adopt the source-kind initial level when it grants more trust.
-                    let stored = row.trust_level;
-                    if !stored.is_active() || stored.severity() <= initial_level.severity() {
-                        stored
-                    } else {
-                        *initial_level
-                    }
-                } else {
-                    row.trust_level
-                }
-            } else {
-                *initial_level
-            };
-            let source_path = meta.skill_dir.to_str();
-            if let Err(e) = memory
-                .sqlite()
-                .upsert_skill_trust(
-                    &meta.name,
-                    trust_level,
-                    source_kind,
-                    None,
-                    source_path,
-                    &current_hash,
-                )
-                .await
-            {
-                tracing::warn!("failed to record trust for '{}': {e:#}", meta.name);
-            }
-        }
-    }
-
-    /// Rebuild or sync the in-memory skill matcher and BM25 index after a registry update.
-    async fn rebuild_skill_matcher(&mut self, all_meta: &[&zeph_skills::loader::SkillMeta]) {
-        let provider = self.embedding_provider.clone();
-        let embed_timeout =
-            std::time::Duration::from_secs(self.runtime.config.timeouts.embedding_seconds);
-        let embed_fn = move |text: &str| -> zeph_skills::matcher::EmbedFuture {
-            let owned = text.to_owned();
-            let p = provider.clone();
-            Box::pin(async move {
-                if let Ok(result) = tokio::time::timeout(embed_timeout, p.embed(&owned)).await {
-                    result
-                } else {
-                    tracing::warn!(
-                        timeout_secs = embed_timeout.as_secs(),
-                        "skill matcher: embedding timed out"
-                    );
-                    Err(zeph_llm::LlmError::Timeout)
-                }
-            })
-        };
-
-        let needs_inmemory_rebuild = !self
-            .services
-            .skill
-            .matcher
-            .as_ref()
-            .is_some_and(SkillMatcherBackend::is_qdrant);
-
-        if needs_inmemory_rebuild {
-            self.services.skill.matcher = SkillMatcher::new(all_meta, embed_fn)
-                .await
-                .map(SkillMatcherBackend::InMemory);
-        } else if let Some(ref mut backend) = self.services.skill.matcher {
-            let _ = self.channel.send_status("syncing skill index...").await;
-            let on_progress: Option<Box<dyn Fn(usize, usize) + Send>> =
-                self.services.session.status_tx.clone().map(
-                    |tx| -> Box<dyn Fn(usize, usize) + Send> {
-                        Box::new(move |completed, total| {
-                            let msg = format!("Syncing skills: {completed}/{total}");
-                            let _ = tx.send(msg);
-                        })
-                    },
-                );
-            if let Err(e) = backend
-                .sync(
-                    all_meta,
-                    &self.services.skill.embedding_model,
-                    embed_fn,
-                    on_progress,
-                )
-                .await
-            {
-                tracing::warn!("failed to sync skill embeddings: {e:#}");
-            }
-        }
-
-        if self.services.skill.hybrid_search {
-            let descs: Vec<&str> = all_meta.iter().map(|m| m.description.as_str()).collect();
-            let _ = self.channel.send_status("rebuilding search index...").await;
-            self.services.skill.rebuild_bm25(&descs);
-        }
-    }
-
-    #[tracing::instrument(name = "core.agent.reload_skills", skip_all, level = "debug")]
-    async fn reload_skills(&mut self) {
-        let old_fp = self.services.skill.fingerprint();
-        let reload_paths = if let Some(ref supplier) = self.services.skill.plugin_dirs_supplier {
-            let plugin_dirs = supplier();
-            let mut paths = self.services.skill.skill_paths.clone();
-            for dir in plugin_dirs {
-                if !paths.contains(&dir) {
-                    paths.push(dir);
-                }
-            }
-            paths
-        } else {
-            self.services.skill.skill_paths.clone()
-        };
-        self.services.skill.registry.write().reload(&reload_paths);
-        if self.services.skill.fingerprint() == old_fp {
-            return;
-        }
-        let _ = self.channel.send_status("reloading skills...").await;
-
-        let all_meta = self
-            .services
-            .skill
-            .registry
-            .read()
-            .all_meta()
-            .into_iter()
-            .cloned()
-            .collect::<Vec<_>>();
-
-        self.update_trust_for_reloaded_skills(&all_meta).await;
-
-        let all_meta_refs = all_meta.iter().collect::<Vec<_>>();
-        self.rebuild_skill_matcher(&all_meta_refs).await;
-
-        let all_skills: Vec<Skill> = {
-            let reg = self.services.skill.registry.read();
-            reg.all_meta()
-                .iter()
-                .filter_map(|m| reg.skill(&m.name).ok())
-                .collect()
-        };
-        let trust_map = self.build_skill_trust_map().await;
-        let empty_health: HashMap<String, (f64, u32)> = HashMap::new();
-        let skills_prompt =
-            state::SkillState::rebuild_prompt(&all_skills, &trust_map, &empty_health);
-        self.services
-            .skill
-            .last_skills_prompt
-            .clone_from(&skills_prompt);
-        let system_prompt = build_system_prompt(&skills_prompt, None);
-        if let Some(msg) = self.msg.messages.first_mut() {
-            msg.content = system_prompt;
-        }
-
-        let _ = self.channel.send_status("").await;
-        tracing::info!(
-            "reloaded {} skill(s)",
-            self.services.skill.registry.read().all_meta().len()
-        );
-    }
-
-    async fn reload_instructions(&mut self) {
-        // Drain any additional queued events before reloading to avoid redundant reloads.
-        if let Some(ref mut rx) = self.runtime.instructions.reload_rx {
-            while rx.try_recv().is_ok() {}
-        }
-        let Some(ref state) = self.runtime.instructions.reload_state else {
-            return;
-        };
-        let base_dir = state.base_dir.clone();
-        let provider_kinds = state.provider_kinds.clone();
-        let explicit_files = state.explicit_files.clone();
-        let auto_detect = state.auto_detect;
-        let new_blocks = crate::instructions::load_instructions_async(
-            base_dir,
-            provider_kinds,
-            explicit_files,
-            auto_detect,
-        )
-        .await;
-        let old_sources: std::collections::HashSet<_> = self
-            .runtime
-            .instructions
-            .blocks
-            .iter()
-            .map(|b| &b.source)
-            .collect();
-        let new_sources: std::collections::HashSet<_> =
-            new_blocks.iter().map(|b| &b.source).collect();
-        for added in new_sources.difference(&old_sources) {
-            tracing::info!(path = %added.display(), "instruction file added");
-        }
-        for removed in old_sources.difference(&new_sources) {
-            tracing::info!(path = %removed.display(), "instruction file removed");
-        }
-        tracing::info!(
-            old_count = self.runtime.instructions.blocks.len(),
-            new_count = new_blocks.len(),
-            "reloaded instruction files"
-        );
-        self.runtime.instructions.blocks = new_blocks;
-    }
-
-    #[allow(clippy::too_many_lines)]
-    fn reload_config(&mut self) {
-        let Some(path) = self.runtime.lifecycle.config_path.clone() else {
-            return;
-        };
-        let Some(config) = self.load_config_with_overlay(&path) else {
-            return;
-        };
-        let budget_tokens = resolve_context_budget(&config, &self.provider);
-        self.runtime.config.security = config.security;
-        self.runtime.config.timeouts = config.timeouts;
-        self.runtime.config.redact_credentials = config.memory.redact_credentials;
-        self.services.memory.persistence.history_limit = config.memory.history_limit;
-        self.services.memory.persistence.recall_limit = config.memory.semantic.recall_limit;
-        self.services.memory.compaction.summarization_threshold =
-            config.memory.summarization_threshold;
-        self.services.skill.max_active_skills = config.skills.max_active_skills.get();
-        self.services.skill.disambiguation_threshold = config.skills.disambiguation_threshold;
-        self.services.skill.min_injection_score = config.skills.min_injection_score;
-        self.services.skill.cosine_weight = config.skills.cosine_weight.clamp(0.0, 1.0);
-        self.services.skill.hybrid_search = config.skills.hybrid_search;
-        {
-            let alpha = config.skills.bm25_alpha;
-            if !(0.0..=1.0).contains(&alpha) {
-                tracing::warn!(
-                    bm25_alpha = alpha,
-                    "bm25_alpha is outside [0.0, 1.0]; clamping to valid range"
-                );
-            }
-            self.services.skill.bm25_alpha = alpha.clamp(0.0, 1.0);
-        }
-        self.services.skill.two_stage_matching = config.skills.two_stage_matching;
-        self.services.skill.confusability_threshold =
-            config.skills.confusability_threshold.clamp(0.0, 1.0);
-        self.services.skill.group_structured = config.skills.group_structured;
-        self.services.skill.support_similarity_threshold =
-            config.skills.support_similarity_threshold;
-        config
-            .skills
-            .query_rewrite_provider
-            .as_str()
-            .clone_into(&mut self.services.skill.query_rewrite_provider_name);
-        config
-            .skills
-            .generation_provider
-            .as_str()
-            .clone_into(&mut self.services.skill.generation_provider_name);
-        config
-            .skills
-            .disambiguate_provider
-            .as_str()
-            .clone_into(&mut self.services.skill.disambiguate_provider_name);
-        self.services.skill.generation_timeout_ms = config.skills.generation_timeout_ms;
-        self.services.skill.semantic_scan = config.skills.semantic_scan;
-        config
-            .skills
-            .semantic_scan_provider
-            .as_str()
-            .clone_into(&mut self.services.skill.semantic_scan_provider);
-        self.services.skill.generation_output_dir =
-            config.skills.generation_output_dir.as_deref().map(|p| {
-                if let Some(stripped) = p.strip_prefix("~/") {
-                    dirs::home_dir()
-                        .map_or_else(|| std::path::PathBuf::from(p), |h| h.join(stripped))
-                } else {
-                    std::path::PathBuf::from(p)
-                }
-            });
-
-        self.context_manager.budget = Some(
-            ContextBudget::new(budget_tokens, 0.20).with_graph_enabled(config.memory.graph.enabled),
-        );
-
-        {
-            let graph_cfg = &config.memory.graph;
-            if graph_cfg.rpe.enabled {
-                // Re-create router only if it doesn't exist yet; preserve state on hot-reload.
-                if self.services.memory.extraction.rpe_router.is_none() {
-                    self.services.memory.extraction.rpe_router =
-                        Some(std::sync::Mutex::new(zeph_memory::RpeRouter::new(
-                            graph_cfg.rpe.threshold,
-                            graph_cfg.rpe.max_skip_turns,
-                        )));
-                }
-            } else {
-                self.services.memory.extraction.rpe_router = None;
-            }
-            self.services.memory.extraction.graph_config = graph_cfg.clone();
-        }
-        self.context_manager.soft_compaction_threshold = config.memory.soft_compaction_threshold;
-        self.context_manager.hard_compaction_threshold = config.memory.hard_compaction_threshold;
-        self.context_manager.compaction_preserve_tail = config.memory.compaction_preserve_tail;
-        self.context_manager
-            .set_compaction_cooldown_turns(config.memory.compaction_cooldown_turns);
-        self.context_manager.prune_protect_tokens = config.memory.prune_protect_tokens;
-        self.context_manager.compression = config.memory.compression.clone();
-        self.context_manager.routing = config.memory.store_routing.clone();
-        // Resolve routing_classifier_provider from the provider pool (#2484).
-        self.context_manager.store_routing_provider = if config
-            .memory
-            .store_routing
-            .routing_classifier_provider
-            .is_empty()
-        {
-            None
-        } else {
-            let resolved = self.resolve_background_provider(
-                config
-                    .memory
-                    .store_routing
-                    .routing_classifier_provider
-                    .as_str(),
-            );
-            Some(std::sync::Arc::new(resolved))
-        };
-        self.services
-            .memory
-            .persistence
-            .cross_session_score_threshold = config.memory.cross_session_score_threshold;
-
-        self.services.index.repo_map_tokens = config.index.repo_map_tokens;
-        self.services.index.repo_map_ttl =
-            std::time::Duration::from_secs(config.index.repo_map_ttl_secs);
-
-        self.services
-            .session
-            .hooks_config
-            .cwd_changed
-            .clone_from(&config.hooks.cwd_changed);
-        self.services
-            .session
-            .hooks_config
-            .permission_denied
-            .clone_from(&config.hooks.permission_denied);
-        self.services
-            .session
-            .hooks_config
-            .turn_complete
-            .clone_from(&config.hooks.turn_complete);
-        // file_changed_hooks require watcher restart to take effect — skipped here.
-
-        tracing::info!("config reloaded");
-    }
-
-    /// Load config from disk, apply plugin overlays, and warn on shell divergence.
-    ///
-    /// Returns `None` when loading or overlay merge fails (caller keeps prior runtime state).
-    fn load_config_with_overlay(&mut self, path: &std::path::Path) -> Option<Config> {
-        let mut config = match Config::load(path) {
-            Ok(c) => c,
-            Err(e) => {
-                tracing::warn!("config reload failed: {e:#}");
-                return None;
-            }
-        };
-
-        // Re-apply plugin overlays. On error, keep previous runtime state intact.
-        let new_overlay = if self.runtime.lifecycle.plugins_dir.as_os_str().is_empty() {
-            None
-        } else {
-            match zeph_plugins::apply_plugin_config_overlays(
-                &mut config,
-                &self.runtime.lifecycle.plugins_dir,
-            ) {
-                Ok(o) => Some(o),
-                Err(e) => {
-                    tracing::warn!(
-                        "plugin overlay merge failed during reload: {e:#}; \
-                         keeping previous runtime state"
-                    );
-                    return None;
-                }
-            }
-        };
-
-        // M4: detect shell-level divergence from the baked-in executor and warn loudly.
-        // ShellExecutor is not rebuilt on hot-reload; only skill threshold is live.
-        // A follow-up P2 issue tracks live-rebuild of ShellExecutor.
-        if let Some(ref overlay) = new_overlay {
-            self.warn_on_shell_overlay_divergence(overlay, &config);
-        }
-        Some(config)
-    }
-
-    /// React to shell policy divergence detected on hot-reload.
-    ///
-    /// `blocked_commands` is rebuilt live via `ShellPolicyHandle::rebuild` — no restart needed.
-    /// `allowed_commands` cannot be rebuilt (feeds sandbox path intersection at construction time)
-    /// — emit a warn + status banner when it changes.
-    fn warn_on_shell_overlay_divergence(
-        &self,
-        new_overlay: &zeph_plugins::ResolvedOverlay,
-        config: &Config,
-    ) {
-        let new_blocked: Vec<String> = {
-            let mut v = config.tools.shell.blocked_commands.clone();
-            v.sort();
-            v
-        };
-        let new_allowed: Vec<String> = {
-            let mut v = config.tools.shell.allowed_commands.clone();
-            v.sort();
-            v
-        };
-
-        let startup = &self.runtime.lifecycle.startup_shell_overlay;
-        let blocked_changed = new_blocked != startup.blocked;
-        let allowed_changed = new_allowed != startup.allowed;
-
-        // blocked_commands IS rebuilt live — emit info-level confirmation only.
-        if blocked_changed && let Some(ref h) = self.runtime.lifecycle.shell_policy_handle {
-            h.rebuild(&config.tools.shell);
-            tracing::info!(
-                blocked_count = h.snapshot_blocked().len(),
-                "shell blocked_commands rebuilt from hot-reload"
-            );
-        }
-
-        // allowed_commands cannot be rebuilt — sandbox path intersection is computed at
-        // executor construction time. Warn loudly so the user restarts.
-        //
-        // Note: when base `allowed_commands` is empty (the default), the overlay's
-        // intersection semantics keep it empty, so this branch is silently unreachable
-        // for users who do not set a non-empty base list.
-        if allowed_changed {
-            let msg = "plugin config overlay changed shell allowed_commands; RESTART REQUIRED \
-                 for sandbox path recomputation (blocked_commands was rebuilt live)";
-            tracing::warn!("{msg}");
-            if let Some(ref tx) = self.services.session.status_tx {
-                let _ = tx.send(msg.to_owned());
-            }
-        }
-
-        let _ = new_overlay;
-    }
-
-    /// Run `SideQuest` tool output eviction pass (#1885).
-    ///
-    /// PERF-1 fix: two-phase non-blocking design.
-    ///
-    /// Phase 1 (apply, this turn): check for a background LLM result spawned last turn,
-    /// validate and apply it immediately.
-    ///
-    /// Phase 2 (schedule, this turn): rebuild cursors and spawn a background `tokio::spawn`
-    /// task for the LLM call. The result is stored in `pending_sidequest_result` and applied
-    /// next turn, so the current agent turn is never blocked by the LLM call.
-    fn maybe_sidequest_eviction(&mut self) {
-        // S1 runtime guard: warn when SideQuest is enabled alongside a non-Reactive pruning
-        // strategy — the two systems share the same pool of evictable tool outputs and can
-        // interfere. Disable sidequest.enabled when pruning_strategy != Reactive.
-        if self.services.sidequest.config.enabled {
-            use crate::config::PruningStrategy;
-            if !matches!(
-                self.context_manager.compression.pruning_strategy,
-                PruningStrategy::Reactive
-            ) {
-                tracing::warn!(
-                    strategy = ?self.context_manager.compression.pruning_strategy,
-                    "sidequest is enabled alongside a non-Reactive pruning strategy; \
-                     consider disabling sidequest.enabled to avoid redundant eviction"
-                );
-            }
-        }
-
-        // Guard: do not evict while a focus session is active.
-        if self.services.focus.is_active() {
-            tracing::debug!("sidequest: skipping — focus session active");
-            // Drop any pending result — cursors may be stale relative to focus truncation.
-            self.services.compression.pending_sidequest_result = None;
-            return;
-        }
-
-        // Phase 1: apply pending result from last turn's background LLM call.
-        self.sidequest_apply_pending();
-
-        // Phase 2: rebuild cursors and schedule the next background eviction LLM call.
-        self.sidequest_schedule_next();
-    }
-
-    fn sidequest_apply_pending(&mut self) {
-        let Some(handle) = self.services.compression.pending_sidequest_result.take() else {
-            return;
-        };
-        // `try_join` is non-blocking: if the task isn't done yet, `Err(handle)` is returned
-        // and we reschedule below.
-        let result = match handle.try_join() {
-            Ok(result) => result,
-            Err(_handle) => {
-                // Task still running — drop it; a fresh one is scheduled below.
-                tracing::debug!("sidequest: background LLM task not yet complete, rescheduling");
-                return;
-            }
-        };
-        match result {
-            Ok(Some(evicted_indices)) if !evicted_indices.is_empty() => {
-                let cursors_snapshot = self.services.sidequest.tool_output_cursors.clone();
-                let freed = self.services.sidequest.apply_eviction(
-                    &mut self.msg.messages,
-                    &evicted_indices,
-                    &self.runtime.metrics.token_counter,
-                );
-                if freed > 0 {
-                    self.recompute_prompt_tokens();
-                    // C1 fix: prevent maybe_compact() from firing in the same turn.
-                    // cooldown=0: eviction does not impose post-compaction cooldown.
-                    self.context_manager.set_compaction_state(
-                        crate::agent::context_manager::CompactionState::CompactedThisTurn {
-                            cooldown: 0,
-                        },
-                    );
-                    tracing::info!(
-                        freed_tokens = freed,
-                        evicted_cursors = evicted_indices.len(),
-                        pass = self.services.sidequest.passes_run,
-                        "sidequest eviction complete"
-                    );
-                    if let Some(ref d) = self.runtime.debug.debug_dumper {
-                        d.dump_sidequest_eviction(&cursors_snapshot, &evicted_indices, freed);
-                    }
-                    if let Some(ref tx) = self.services.session.status_tx {
-                        let _ = tx.send(format!("SideQuest evicted {freed} tokens"));
-                    }
-                } else {
-                    // apply_eviction returned 0 — clear spinner so it doesn't dangle.
-                    if let Some(ref tx) = self.services.session.status_tx {
-                        let _ = tx.send(String::new());
-                    }
-                }
-            }
-            Ok(None | Some(_)) => {
-                tracing::debug!("sidequest: pending result: no cursors to evict");
-                if let Some(ref tx) = self.services.session.status_tx {
-                    let _ = tx.send(String::new());
-                }
-            }
-            Err(e) => {
-                tracing::debug!("sidequest: background task error: {e}");
-                if let Some(ref tx) = self.services.session.status_tx {
-                    let _ = tx.send(String::new());
-                }
-            }
-        }
-    }
-
-    fn sidequest_schedule_next(&mut self) {
-        use zeph_llm::provider::{Message, MessageMetadata, Role};
-
-        self.services
-            .sidequest
-            .rebuild_cursors(&self.msg.messages, &self.runtime.metrics.token_counter);
-
-        if self.services.sidequest.tool_output_cursors.is_empty() {
-            tracing::debug!("sidequest: no eligible cursors");
-            return;
-        }
-
-        let prompt = self.services.sidequest.build_eviction_prompt();
-        let max_eviction_ratio = self.services.sidequest.config.max_eviction_ratio;
-        let n_cursors = self.services.sidequest.tool_output_cursors.len();
-        // Clone the provider so the spawn closure owns it without borrowing self.
-        let provider = self.summary_or_primary_provider().clone();
-
-        let eviction_future = async move {
-            let msgs = [Message {
-                role: Role::User,
-                content: prompt,
-                parts: vec![],
-                metadata: MessageMetadata::default(),
-            }];
-            let response =
-                match tokio::time::timeout(std::time::Duration::from_secs(5), provider.chat(&msgs))
-                    .await
-                {
-                    Ok(Ok(r)) => r,
-                    Ok(Err(e)) => {
-                        tracing::debug!("sidequest bg: LLM call failed: {e:#}");
-                        return None;
-                    }
-                    Err(_) => {
-                        tracing::debug!("sidequest bg: LLM call timed out");
-                        return None;
-                    }
-                };
-
-            let start = response.find('{')?;
-            let end = response.rfind('}')?;
-            if start > end {
-                return None;
-            }
-            let json_slice = &response[start..=end];
-            let parsed: sidequest::EvictionResponse = serde_json::from_str(json_slice).ok()?;
-            let mut valid: Vec<usize> = parsed
-                .del_cursors
-                .into_iter()
-                .filter(|&c| c < n_cursors)
-                .collect();
-            valid.sort_unstable();
-            valid.dedup();
-            #[allow(
-                clippy::cast_precision_loss,
-                clippy::cast_possible_truncation,
-                clippy::cast_sign_loss
-            )]
-            let max_evict = ((n_cursors as f32) * max_eviction_ratio).ceil() as usize;
-            valid.truncate(max_evict);
-            Some(valid)
-        };
-        let handle = self.runtime.lifecycle.task_supervisor.spawn_oneshot(
-            std::sync::Arc::from("agent.sidequest.eviction"),
-            move || eviction_future,
-        );
-        self.services.compression.pending_sidequest_result = Some(handle);
-        tracing::debug!("sidequest: background LLM eviction task spawned");
-        if let Some(ref tx) = self.services.session.status_tx {
-            let _ = tx.send("SideQuest: scoring tool outputs...".into());
-        }
-    }
-
-    /// Return an `McpDispatch` adapter backed by the agent's MCP manager, if present.
-    fn mcp_dispatch(&self) -> Option<McpManagerDispatch> {
-        self.services
-            .mcp
-            .manager
-            .as_ref()
-            .map(|m| McpManagerDispatch(Arc::clone(m)))
-    }
-
-    /// Check if the process cwd has changed since last call and fire `CwdChanged` hooks.
-    ///
-    /// Called after each tool batch completes. The check is a single syscall and has
-    /// negligible cost. Only fires when cwd actually changed (defense-in-depth: normally
-    /// only `set_working_directory` changes cwd; shell child processes cannot affect it).
-    pub(crate) async fn check_cwd_changed(&mut self) {
-        let current = match std::env::current_dir() {
-            Ok(p) => p,
-            Err(e) => {
-                tracing::warn!("check_cwd_changed: failed to get cwd: {e}");
-                return;
-            }
-        };
-        if current == self.runtime.lifecycle.last_known_cwd {
-            return;
-        }
-        let old_cwd =
-            std::mem::replace(&mut self.runtime.lifecycle.last_known_cwd, current.clone());
-        self.services.session.env_context.working_dir = current.display().to_string();
-
-        tracing::info!(
-            old = %old_cwd.display(),
-            new = %current.display(),
-            "working directory changed"
-        );
-
-        let _ = self
-            .channel
-            .send_status("Working directory changed\u{2026}")
-            .await;
-
-        let hooks = self.services.session.hooks_config.cwd_changed.clone();
-        if hooks.is_empty() {
-            tracing::debug!("CwdChanged: no hooks configured, skipping");
-        } else {
-            tracing::debug!(count = hooks.len(), "CwdChanged: firing hooks");
-            let mut env = std::collections::HashMap::new();
-            env.insert("ZEPH_OLD_CWD".to_owned(), old_cwd.display().to_string());
-            env.insert("ZEPH_NEW_CWD".to_owned(), current.display().to_string());
-            let dispatch = self.mcp_dispatch();
-            let mcp: Option<&dyn zeph_subagent::McpDispatch> = dispatch
-                .as_ref()
-                .map(|d| d as &dyn zeph_subagent::McpDispatch);
-            if let Err(e) = zeph_subagent::hooks::fire_hooks(&hooks, &env, mcp, None).await {
-                tracing::warn!(error = %e, "CwdChanged hook failed");
-            } else {
-                tracing::info!(count = hooks.len(), "CwdChanged: hooks fired");
-            }
-        }
-
-        let _ = self.channel.send_status("").await;
-    }
-
-    /// Handle a `FileChangedEvent` from the file watcher.
-    pub(crate) async fn handle_file_changed(
-        &mut self,
-        event: crate::file_watcher::FileChangedEvent,
-    ) {
-        tracing::info!(path = %event.path.display(), "file changed");
-
-        let _ = self
-            .channel
-            .send_status("Running file-change hook\u{2026}")
-            .await;
-
-        let hooks = self
-            .services
-            .session
-            .hooks_config
-            .file_changed_hooks
-            .clone();
-        if hooks.is_empty() {
-            tracing::debug!(path = %event.path.display(), "FileChanged: no hooks configured, skipping");
-        } else {
-            tracing::debug!(count = hooks.len(), path = %event.path.display(), "FileChanged: firing hooks");
-            let mut env = std::collections::HashMap::new();
-            env.insert(
-                "ZEPH_CHANGED_PATH".to_owned(),
-                event.path.display().to_string(),
-            );
-            let dispatch = self.mcp_dispatch();
-            let mcp: Option<&dyn zeph_subagent::McpDispatch> = dispatch
-                .as_ref()
-                .map(|d| d as &dyn zeph_subagent::McpDispatch);
-            if let Err(e) = zeph_subagent::hooks::fire_hooks(&hooks, &env, mcp, None).await {
-                tracing::warn!(error = %e, "FileChanged hook failed");
-            } else {
-                tracing::info!(count = hooks.len(), path = %event.path.display(), "FileChanged: hooks fired");
-            }
-        }
-
-        let _ = self.channel.send_status("").await;
-    }
-
     /// If the compression spectrum is enabled and a promotion engine is wired, spawn a
     /// background scan task.
     ///
@@ -3758,253 +1884,6 @@ impl<C: Channel> Agent<C> {
         if accepted {
             tracing::debug!("compression_spectrum: promotion scan task enqueued");
         }
-    }
-}
-/// Estimates the JSON payload size of a single [`zeph_llm::provider::Message`] for token-budget
-/// accounting.
-///
-/// When `parts` is empty the message is a legacy text-only message and `content.len()` is used
-/// directly. Otherwise each part is measured individually so that structured variants (images,
-/// tool invocations, thinking blocks) are accounted for rather than relying on the already-flat
-/// `content` string, which may not reflect the actual API payload size.
-pub(crate) fn estimate_parts_size(m: &zeph_llm::provider::Message) -> usize {
-    use zeph_llm::provider::MessagePart;
-    if m.parts.is_empty() {
-        return m.content.len();
-    }
-    m.parts
-        .iter()
-        .map(|p| match p {
-            MessagePart::Text { text }
-            | MessagePart::Recall { text }
-            | MessagePart::CodeContext { text }
-            | MessagePart::Summary { text }
-            | MessagePart::CrossSession { text } => text.len(),
-            MessagePart::ToolOutput { body, .. } => body.len(),
-            MessagePart::ToolUse { id, name, input } => {
-                50 + id.len() + name.len() + input.to_string().len()
-            }
-            MessagePart::ToolResult {
-                tool_use_id,
-                content,
-                ..
-            } => 50 + tool_use_id.len() + content.len(),
-            MessagePart::Image(img) => img.data.len() * 4 / 3,
-            MessagePart::ThinkingBlock {
-                thinking,
-                signature,
-            } => 50 + thinking.len() + signature.len(),
-            MessagePart::RedactedThinkingBlock { data } => data.len(),
-            MessagePart::Compaction { summary } => summary.len(),
-            _ => 0,
-        })
-        .sum()
-}
-
-/// Applies token-budget truncation and orphaned-tool-pair pruning to a parent message slice.
-///
-/// Budget truncation keeps the **most recent** messages that fit within `max_chars`
-/// (a suffix), so the subagent always receives the freshest context.
-///
-/// Two passes are performed after budget truncation:
-///
-/// 1. Remove `ToolResult` parts from user messages whose matching `ToolUse` is no longer in the
-///    slice (truncated away).
-/// 2. Remove `ToolUse` parts from **interior** assistant messages whose matching `ToolResult`
-///    was removed in pass 1 or was already absent. The trailing assistant message is exempt —
-///    its unanswered `ToolUse` calls are not orphaned; the slice just ends before the result.
-///
-/// Messages that become fully empty after pruning are removed from `msgs`.
-///
-/// `rebuild_content` is called **only** when `retain` actually removed parts — preserving the
-/// existing `content` field (and any `ThinkingBlock` text embedded there) for unmodified
-/// messages.
-pub(crate) fn trim_parent_messages(msgs: &mut Vec<zeph_llm::provider::Message>, max_chars: usize) {
-    use zeph_llm::provider::{MessagePart, Role};
-
-    // Token-budget cap: keep the most recent messages that fit within max_chars.
-    // We iterate from the end (newest) and drain from the front once the budget is exceeded,
-    // so the subagent always receives the most recent context rather than stale early messages.
-    let mut total_chars = 0usize;
-    let mut drop_before = 0usize; // index of the first message to keep
-    for (i, m) in msgs.iter().enumerate().rev() {
-        total_chars += estimate_parts_size(m);
-        if total_chars > max_chars {
-            drop_before = i + 1;
-            break;
-        }
-    }
-    if drop_before > 0 {
-        msgs.drain(..drop_before);
-    }
-
-    // Pass 1: collect ToolUse IDs emitted by assistant messages; prune orphaned ToolResult
-    // parts from user messages that reference a ToolUse no longer present in the slice.
-    // Use owned Strings to avoid holding immutable borrows across the subsequent mutable loop.
-    let emitted_tool_ids: std::collections::HashSet<String> = msgs
-        .iter()
-        .filter(|m| m.role == Role::Assistant)
-        .flat_map(|m| m.parts.iter())
-        .filter_map(|p| {
-            if let MessagePart::ToolUse { id, .. } = p {
-                Some(id.clone())
-            } else {
-                None
-            }
-        })
-        .collect();
-
-    let mut orphans_removed = 0usize;
-    for m in msgs.iter_mut() {
-        if m.role != Role::User || m.parts.is_empty() {
-            continue;
-        }
-        let before = m.parts.len();
-        m.parts.retain(|p| match p {
-            MessagePart::ToolResult { tool_use_id, .. } => {
-                emitted_tool_ids.contains(tool_use_id.as_str())
-            }
-            _ => true,
-        });
-        let dropped = before - m.parts.len();
-        if dropped > 0 {
-            orphans_removed += dropped;
-            if m.parts.is_empty() {
-                m.content.clear();
-            } else {
-                m.rebuild_content();
-            }
-        }
-    }
-
-    // Pass 2: collect ToolResult IDs present in user messages after pass 1; prune ToolUse
-    // parts from assistant messages whose result is confirmed absent.
-    //
-    // The trailing assistant message is exempt: it may legitimately contain unanswered
-    // ToolUse calls (the slice ends before the result arrives). Only interior assistant
-    // messages — those followed by at least one user message — can have provably orphaned
-    // ToolUse parts (the conversation moved on without answering them).
-    let consumed_tool_ids: std::collections::HashSet<String> = msgs
-        .iter()
-        .filter(|m| m.role == Role::User)
-        .flat_map(|m| m.parts.iter())
-        .filter_map(|p| {
-            if let MessagePart::ToolResult { tool_use_id, .. } = p {
-                Some(tool_use_id.clone())
-            } else {
-                None
-            }
-        })
-        .collect();
-
-    // Index of the last assistant message — exempt from pass 2.
-    let last_assistant_idx = msgs
-        .iter()
-        .enumerate()
-        .rev()
-        .find(|(_, m)| m.role == Role::Assistant)
-        .map(|(i, _)| i);
-
-    for (idx, m) in msgs.iter_mut().enumerate() {
-        if m.role != Role::Assistant || m.parts.is_empty() {
-            continue;
-        }
-        // Skip the trailing assistant message — its unanswered ToolUse calls are not orphaned.
-        if Some(idx) == last_assistant_idx {
-            continue;
-        }
-        let before = m.parts.len();
-        m.parts.retain(|p| match p {
-            MessagePart::ToolUse { id, .. } => consumed_tool_ids.contains(id.as_str()),
-            _ => true,
-        });
-        let dropped = before - m.parts.len();
-        if dropped > 0 {
-            orphans_removed += dropped;
-            if m.parts.is_empty() {
-                m.content.clear();
-            } else {
-                m.rebuild_content();
-            }
-        }
-    }
-
-    // Remove messages that were emptied by orphan pruning.
-    msgs.retain(|m| !m.content.is_empty() || !m.parts.is_empty());
-
-    if orphans_removed > 0 {
-        tracing::debug!(
-            orphans = orphans_removed,
-            "[subagent] pruned orphaned ToolUse/ToolResult parts from parent context boundary"
-        );
-    }
-}
-
-/// Sanitize text parts of `msgs` through the IPI pipeline.
-///
-/// Only [`MessagePart::Text`] parts are passed through the sanitizer; structured parts
-/// (`ToolUse`, `ToolResult`, `Recall`, `CodeContext`) are left untouched.  After sanitization
-/// the message `content` field is rebuilt to stay consistent with the updated parts.
-fn sanitize_parent_messages(
-    mut msgs: Vec<zeph_llm::provider::Message>,
-    sanitizer: &zeph_sanitizer::ContentSanitizer,
-    source: &zeph_sanitizer::ContentSource,
-) -> Vec<zeph_llm::provider::Message> {
-    use zeph_llm::provider::MessagePart;
-    for msg in &mut msgs {
-        let mut changed = false;
-        for part in &mut msg.parts {
-            if let MessagePart::Text { text } = part {
-                let clean = sanitizer.sanitize(text, source.clone());
-                if clean.body != *text {
-                    *text = clean.body;
-                    changed = true;
-                }
-            }
-        }
-        if changed {
-            msg.rebuild_content();
-        }
-    }
-    msgs
-}
-
-/// Thin wrapper that implements [`zeph_subagent::McpDispatch`] over an [`Arc<zeph_mcp::McpManager>`].
-///
-/// Used to pass MCP tool dispatch capability into `fire_hooks` without coupling
-/// `zeph-subagent` to `zeph-mcp`.
-struct McpManagerDispatch(Arc<zeph_mcp::McpManager>);
-
-impl zeph_subagent::McpDispatch for McpManagerDispatch {
-    fn call_tool<'a>(
-        &'a self,
-        server: &'a str,
-        tool: &'a str,
-        args: serde_json::Value,
-    ) -> std::pin::Pin<
-        Box<dyn std::future::Future<Output = Result<serde_json::Value, String>> + Send + 'a>,
-    > {
-        Box::pin(async move {
-            self.0
-                .call_tool(server, tool, args)
-                .await
-                .map(|result| {
-                    // Extract text content from the MCP response as a JSON value.
-                    let texts: Vec<serde_json::Value> = result
-                        .content
-                        .iter()
-                        .filter_map(|c| {
-                            if let rmcp::model::RawContent::Text(t) = &c.raw {
-                                Some(serde_json::Value::String(t.text.clone()))
-                            } else {
-                                None
-                            }
-                        })
-                        .collect();
-                    serde_json::Value::Array(texts)
-                })
-                .map_err(|e| e.to_string())
-        })
     }
 }
